Moved MSParticleType and TransportalizerParticle to new particles package, extracted particle provider subscribe event to ClientProxy, added javadoc comment to MSParticleType
