+++ src/main/java/com/mraof/minestuck/particles/TransportalizerParticle.java
@@ -1,10 +1,8 @@
-package com.mraof.minestuck.client.particle;
+package com.mraof.minestuck.particles;
 
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.particle.*;
 import net.minecraft.core.particles.SimpleParticleType;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 
 public class TransportalizerParticle extends TextureSheetParticle
 {
@@ -33,7 +31,6 @@ public class TransportalizerParticle extends TextureSheetParticle
 		setSpriteFromAge(this.spriteSet);
 	}
 	
-	@OnlyIn(Dist.CLIENT)
 	public static class Provider implements ParticleProvider<SimpleParticleType>
 	{
 		private final SpriteSet spriteSet;
